+++ pga/src/main/java/unlu/sip/pga/repositories/EjercicioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import unlu.sip.pga.entities.Ejercicio;
 import java.util.List;
 
-public interface EjercicioRepository extends JpaRepository<Ejercicio, Long> {
-    List<Ejercicio> findByModuloId(Long idModulo);
+public interface EjercicioRepository extends JpaRepository<Ejercicio, Integer> {
+    List<Ejercicio> findByModuloId(Integer idModulo);
 }
