+++ pga/src/main/java/unlu/sip/pga/repositories/SuscripcionRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import unlu.sip.pga.entities.Suscripcion;
 import java.util.List;
 
-public interface SuscripcionRepository extends JpaRepository<Suscripcion, Long> {
-    List<Suscripcion> findByUsuarioId(Long idUsuario);
+public interface SuscripcionRepository extends JpaRepository<Suscripcion, Integer> {
+    List<Suscripcion> findByUsuarioId(Integer idUsuario);
 }
